backend/requirements.txt
aiosqlite
fastapi
httpx
nashpy
numpy
pydantic
python-dotenv
python-multipart
scipy
sse-starlette
uvicorn
